+++ common/addons/chunk-generator-noise-3d/src/main/java/com/dfsek/terra/addons/chunkgenerator/config/NoiseChunkGeneratorPackConfigTemplate.java
@@ -4,7 +4,7 @@ import com.dfsek.tectonic.api.config.template.ConfigTemplate;
 import com.dfsek.tectonic.api.config.template.annotations.Default;
 import com.dfsek.tectonic.api.config.template.annotations.Value;
 
-import com.dfsek.terra.addons.chunkgenerator.palette.slant.SlantHolder;
+import com.dfsek.terra.addons.chunkgenerator.generation.math.SlantCalculationMethod;
 import com.dfsek.terra.api.config.meta.Meta;
 import com.dfsek.terra.api.properties.Properties;
 
@@ -24,7 +24,7 @@ public class NoiseChunkGeneratorPackConfigTemplate implements ConfigTemplate, Pr
 
     @Value("slant.calculation-method")
     @Default
-    private SlantHolder.@Meta CalculationMethod slantCalculationMethod = SlantHolder.CalculationMethod.Derivative;
+    private @Meta SlantCalculationMethod slantCalculationMethod = SlantCalculationMethod.Derivative;
 
     public int getElevationBlend() {
         return elevationBlend;
@@ -38,7 +38,7 @@ public class NoiseChunkGeneratorPackConfigTemplate implements ConfigTemplate, Pr
         return verticalRes;
     }
 
-    public SlantHolder.CalculationMethod getSlantCalculationMethod() {
+    public SlantCalculationMethod getSlantCalculationMethod() {
         return slantCalculationMethod;
     }
 }
