+++ common/addons/chunk-generator-noise-3d/src/main/java/com/dfsek/terra/addons/chunkgenerator/palette/slant/SingleSlantHolder.java
@@ -1,5 +1,6 @@
 package com.dfsek.terra.addons.chunkgenerator.palette.slant;
 
+import com.dfsek.terra.addons.chunkgenerator.generation.math.SlantCalculationMethod;
 import com.dfsek.terra.addons.chunkgenerator.palette.PaletteHolder;
 
 
@@ -7,7 +8,7 @@ final class SingleSlantHolder extends SlantHolderImpl {
 
     private final SlantHolder.Layer layer;
 
-    public SingleSlantHolder(SlantHolder.Layer layer, int slantDepth, CalculationMethod calculationMethod) {
+    public SingleSlantHolder(SlantHolder.Layer layer, int slantDepth, SlantCalculationMethod calculationMethod) {
         super(slantDepth, calculationMethod);
         this.layer = layer;
     }
