+++ common/addons/chunk-generator-noise-3d/src/main/java/com/dfsek/terra/addons/chunkgenerator/palette/slant/SlantHolderImpl.java
@@ -1,25 +1,18 @@
 package com.dfsek.terra.addons.chunkgenerator.palette.slant;
 
-import com.dfsek.terra.addons.chunkgenerator.generation.math.samplers.Sampler3D;
+import com.dfsek.terra.addons.chunkgenerator.generation.math.SlantCalculationMethod;
 
 
 public abstract class SlantHolderImpl implements SlantHolder {
     protected final boolean floorToThreshold;
-    private final SlantHolder.CalculationMethod calculationMethod;
     private final int slantDepth;
 
-    protected SlantHolderImpl(int slantDepth, CalculationMethod calculationMethod) {
+    protected SlantHolderImpl(int slantDepth, SlantCalculationMethod calculationMethod) {
         this.floorToThreshold = calculationMethod.floorToThreshold();
-        this.calculationMethod = calculationMethod;
         this.slantDepth = slantDepth;
     }
 
     protected abstract double getSlantThreshold();
-
-    @Override
-    public final double calculateSlant(Sampler3D sampler, double x, double y, double z) {
-        return calculationMethod.slant(sampler, x, y, z);
-    }
 
     @Override
     public final boolean isAboveDepth(int depth) {
